pyproject.toml
[project]
name = "labgchainv0"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "instructor>=1.11.3",
    "langchain>=0.3.27",
    "langchain-openai>=0.3.33",
    "openai>=1.107.3",
    "portkey-ai>=1.14.4",
    "pydantic>=2.11.9",
    "python-dotenv>=1.1.1",
]
